get the weather using location

+++ app/src/main/java/com/example/weatherforecastapp/home/model/Rain.kt
@@ -1,5 +1,5 @@
 package com.example.weatherforecastapp.home.model
 
 data class Rain(
-    val `1h`: Int
+    val `1h`: Double
 )
